:construction: WiP navigation content
Expanding the design of navigation, and starting out with a basic view of
history. This doesn't allow using history yet, and doesn't even really show
where we are in history yet, but it helps inform where I'll take it.

But it's Sunday morning, I've been tinkering, now it's time to go enjoy the
day...