:sparkles: Flesh out the history list a little more
Also add a simple form of selecting something from history. This will need
improving as it's just a placeholder for the moment. For one thing, when you
visit somewhere from history a fresh copy goes to the top of history, which
isn't what you'd expect.

I also think that I should collapse OpenFile and OpenURL into one message,
at this point, called OpenLocation or similar. And then I should probably
have ReopenLocation that will do similar but not add to history.

I think I might also have to change Viewer so that it either doesn't use a
reactive to specify the location to visit, but instead uses a method -- or
keep it and add a goto_history method that takes a history ID and changes
things around under the hood (actually this might be the better approach).