:sparkles: Add source property to the viewer widget
This adds the ability to access the source of the Markdown being viewed.

Really Textual's Markdown widget should have this; it has access to it, it
holds it internally, but on the couple of occasions where I proposed making
it available via a public API the idea was shot down. For reasons.